pyproject.toml
[project]
name = "immigration-mcp-server"
version = "0.1.0"
description = "MCP Server for US Immigration Law Navigation and SOC/Wage Discovery"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
